update libsecpt to secp256k1-zkp 4f7a50e173f85113a3c495fba053053fc284dcf6

+++ src/secp256k1/src/modules/rangeproof/borromean.h
@@ -15,10 +15,10 @@
 #include "ecmult_gen.h"
 
 int secp256k1_borromean_verify(const secp256k1_ecmult_context* ecmult_ctx, secp256k1_scalar *evalues, const unsigned char *e0, const secp256k1_scalar *s,
- const secp256k1_gej *pubs, const int *rsizes, int nrings, const unsigned char *m, int mlen);
+ const secp256k1_gej *pubs, const size_t *rsizes, size_t nrings, const unsigned char *m, size_t mlen);
 
 int secp256k1_borromean_sign(const secp256k1_ecmult_context* ecmult_ctx, const secp256k1_ecmult_gen_context *ecmult_gen_ctx,
  unsigned char *e0, secp256k1_scalar *s, const secp256k1_gej *pubs, const secp256k1_scalar *k, const secp256k1_scalar *sec,
- const int *rsizes, const int *secidx, int nrings, const unsigned char *m, int mlen);
+ const size_t *rsizes, const size_t *secidx, size_t nrings, const unsigned char *m, size_t mlen);
 
 #endif
